# docs/swagger.yaml
basePath: /
definitions:
  handler.ResponseSuccess:
    properties:
      data: {}
      message:
        type: string
    type: object
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.errorResponseData:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  model.CourseFull:
    properties:
      body:
        type: string
      body_ru:
        type: string
      date:
        type: string
      duration:
        type: string
      format:
        type: string
      price:
        type: string
      term:
        type: string
      title:
        type: string
      title_ru:
        type: string
    type: object
...
  model.TableFull:
    properties:
      date:
        type: string
      duration:
        type: string
      format:
        type: string
      id:
        type: integer
      post_body:
        type: string
      post_body_ru:
        type: string
      post_date:
        type: string
      post_img_path:
        type: string
      post_img_url:
        type: string
      post_title:
        type: string
      post_title_ru:
        type: string
      price:
        type: string
    type: object
  model.TablePost:
    properties:
      date:
        type: string
      duration:
        type: string
      format:
        type: string
      post_body:
        type: string
      post_body_ru:
        type: string
      post_img_url:
        type: string
      post_title:
        type: string
      post_title_ru:
        type: string
      price:
        type: string
    type: object
...
  model.allTable:
    properties:
      allHome:
        items:
          $ref: '#/definitions/model.TableFull'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for Norbekov App
  title: Norbekov API
  version: "1.0"
paths:
...
  /api/table/course/create:
    post:
      consumes:
      - application/json
      description: create course_post
      operationId: create-course_post
      parameters:
      - description: course info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CourseFull'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create CoursePost
      tags:
      - Table
  /api/table/course/delete:
    delete:
      consumes:
      - application/json
      description: delete table
      operationId: delete-course
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TableFull'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Delete Coures
      tags:
      - Table
  /api/table/create:
    post:
      consumes:
      - application/json
      description: create table_post
      operationId: create-table_post
      parameters:
      - description: table info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.TablePost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create TablePost
      tags:
      - Table
  /api/table/delete:
    delete:
      consumes:
      - application/json
      description: delete table
      operationId: delete-table
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TableFull'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Delete table
      tags:
      - Table
  /api/table/update/{id}:
    put:
      consumes:
      - application/json
      description: Update Table by id
      operationId: update-table-id
      parameters:
      - description: Param ID
        in: path
        name: id
        required: true
        type: integer
      - description: home info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.TablePost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update  Table By ID
      tags:
      - Table
  /api/table/upload-img/{id}:
    patch:
      consumes:
      - application/json
      - multipart/form-data
      description: Upload Table Image
      operationId: upload_img_table
      parameters:
      - description: Param ID
        in: path
        name: id
        required: true
        type: integer
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      - application/octet-stream
      - image/png
      - image/jpeg
      - image/jpg
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Upload Table Image
      tags:
      - Table
...
  /table/get:
    get:
      consumes:
      - application/json
      description: get table post
      operationId: get-table
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TableFull'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get table
      tags:
      - Table
  /tables/course/get:
    get:
      consumes:
      - application/json
      description: get  courses
      operationId: get-course
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get courses
      tags:
      - Table
  /tables/get:
    get:
      consumes:
      - application/json
      description: get  tables
      operationId: get-tables
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.allTable'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get tables
      tags:
      - Table
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"